Include project config.nsdepcop only if it exists
This enables skipping CoreCompile in cases where config.nsdepcop is not next to the project, e. g. because it is defined at a solution level and is manually included as AdditionalFile into all projects.
The analyzer itself will still issue a diagnostic in case it does not have any config.nsdepcop